requirements.txt
streamlit>=1.36
pandas==2.2.2
numpy>=1.26,<2.0
scikit-learn>=1.4,<1.6
tensorflow-cpu==2.16.1
yfinance>=0.2.52
matplotlib>=3.8
plotly>=5.22
python-dotenv>=1.0
protobuf<5.0.0
typing-extensions>=4.5.0
